+++ odbc/odbc_types.h
@@ -46,6 +46,7 @@ constexpr const char* SQLSTATE_CONNECTION_TIMEOUT = "HYT01";
 constexpr const char* SQLSTATE_CONNECTION_NOT_OPEN = "08003";
 constexpr const char* SQLSTATE_INVALID_TRANSACTION_OPERATION = "HY012";
 constexpr const char* SQLSTATE_OPTIONAL_FEATURE_NOT_IMPLEMENTED = "HYC00";
+constexpr const char* SQLSTATE_INVALID_DRIVER_COMPLETION = "HY110";
 
 // Note: All SQL_DIAG_* constants are already defined in system ODBC headers (sql.h, sqlext.h)
 // No need to redefine them here